Updated the way in which scripts are minified. Updated minified scripts.
Untested. The goal is to update the minifier plugin and prepare for updating flot and jquery
to newer or newest versions.

Used a different plugin to minify the wrapped JavaScript scripts. The plugin
is modern and maintained.

Added antrun tasks and copy resources to create a short procedure for adding
or updating scripts that'll be minified. The minified scripts end up in the
*/client/resources directory just like before.