Fixed causing an HTTP request to https://host/null
The value of image.src was being set to null, upon drawing the image,
the browser would request "null". Fixed by only drawing the image if
the imageurl variable is defined, not null, and not empty.

I can't find a mroe recent version of this plugin. It may not be needed
with newer version of flot. Will evaluate later.